TB-1462 - DDFileLogger.m line 984 [Crashlytics #287, #302, #311, #320, #326, #328] (maybe)
Copy filePath before using it in the xattr calls.  There is a chance
that renameFile is called on a different thread during this call,
which would change filePath and therefore free the buffer that the
call is using.  I actually can't find a situation where this happens,
since my repro case has gone away again, but this is a possibility
at least.

Switch from using UTF8String to fileSystemRepresentation for this
buffer.  I don't know if there's any difference between these two
calls for these filenames, but it's semantically better to be
using fileSystemRepresentation.