Set NSFileProtectionKey on the temporary file created during compression.
This uses the same protection for the temporary file as is used on
the log file itself.  This means that it has the same security
characteristics.  Also, if the app can run in the background, this
means that it gets NSFileProtectionCompleteUntilFirstUserAuthentication
so that we can do this compression even with the device locked.

This means that in effect we pick up the same decision as used in
DDFileLogger around doesAppRunInBackground.