Fix log rotation with a locked screen on iOS.
If the log file does not exist when we try to open it for writing,
and we're on iOS, and the app runs in the background, then the
problem might be that the screen (and therefore the protected data
in the filesystem) was locked when currentLogFileInfo tried to
create the file.  In this situation, clear currentLogFileInfo and retry.

To further mitigate this, refuse to rotate the log file at all when
the screen is locked.  This means that we'll keep our current file
handle open, which we can continue to write to.